requirements.txt
# Core ML / retrieval
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
scikit-learn>=1.4.0
joblib>=1.4.0

# Deep learning
torch>=2.2.0
transformers>=4.40.0
bitsandbytes>=0.43.0

# Data
pandas>=2.2.0
numpy>=1.26.0

# Generation APIs
openai>=1.30.0
google-generativeai>=0.7.0

# Utilities
tqdm>=4.66.0
